buncha new scripts

+++ .config/swayimg/init.lua
@@ -8,20 +8,20 @@ swayimg.imagelist.enable_fsmon(true) -- auto-refresh on file changes
 local screenshots_dir = os.getenv("HOME") .. "/Pictures/Screenshots"
 local last_dir = nil
 swayimg.gallery.on_image_change(function()
-	local image = swayimg.gallery.get_image()
-	if image and image.path then
-		local dir = image.path:match("^(.*)/") or ""
-		if dir == screenshots_dir then
-			swayimg.imagelist.enable_reverse(true)
-			-- Jump to first (newest) when entering Screenshots directory
-			if last_dir ~= screenshots_dir then
-				swayimg.gallery.switch_image("first")
-			end
-		else
-			swayimg.imagelist.enable_reverse(false)
-		end
-		last_dir = dir
-	end
+    local image = swayimg.gallery.get_image()
+    if image and image.path then
+        local dir = image.path:match("^(.*)/") or ""
+        if dir == screenshots_dir then
+            swayimg.imagelist.enable_reverse(true)
+            -- Jump to first (newest) when entering Screenshots directory
+            if last_dir ~= screenshots_dir then
+                swayimg.gallery.switch_image("first")
+            end
+        else
+            swayimg.imagelist.enable_reverse(false)
+        end
+        last_dir = dir
+    end
 end)
 
 -- ============================================================================
@@ -55,27 +55,38 @@ swayimg.gallery.set_unselected_color(0x00000000)
 
 -- Copy image path to clipboard
 swayimg.gallery.on_key("Ctrl+c", function()
-	local image = swayimg.gallery.get_image()
-	os.execute(string.format("wl-copy '%s' && notify-send 'Copied image path' -i '%s'", image.path, image.path))
+    local image = swayimg.gallery.get_image()
+    os.execute(string.format("wl-copy '%s' && notify-send 'Copied image path' -i '%s'", image.path, image.path))
 end)
 
 -- Add image to Favorites
 swayimg.gallery.on_key("Ctrl+f", function()
-	local image = swayimg.gallery.get_image()
-	os.execute(
-		string.format(
-			"cp '%s' ~/Pictures/Favorites/ && notify-send 'Added image to Favorites' -i '%s'",
-			image.path,
-			image.path
-		)
-	)
+    local image = swayimg.gallery.get_image()
+    os.execute(
+        string.format(
+            "cp '%s' ~/Pictures/Favorites/ && notify-send 'Added image to Favorites' -i '%s'",
+            image.path,
+            image.path
+        )
+    )
 end)
+
+-- function to confirm, feel free to reuse, only used on delete
+local function confirm(prompt)
+    local handle =
+        io.popen("printf 'Yes\\nNo' | wofi -c ~/.config/wofi/yes_or_no.conf --dmenu --prompt '" .. prompt .. "'")
+    local result = handle:read("*l")
+    handle:close()
+    return result == "Yes"
+end
 
 -- Delete image
 swayimg.gallery.on_key("Delete", function()
-	local image = swayimg.gallery.get_image()
-	os.execute(string.format("rm '%s' && notify-send 'Image deleted' -i '%s'", image.path, image.path))
-	swayimg.imagelist.remove(image.path)
+    local image = swayimg.gallery.get_image()
+    if confirm("Delete this image?") then
+        os.execute(string.format("rm '%s' && notify-send 'Image deleted' -i '%s'", image.path, image.path))
+        swayimg.imagelist.remove(image.path)
+    end
 end)
 
 -- ============================================================================
@@ -93,37 +104,39 @@ swayimg.viewer.set_window_background(0x00000000)
 
 -- Copy image path to clipboard
 swayimg.viewer.on_key("Ctrl+c", function()
-	local image = swayimg.viewer.get_image()
-	os.execute(string.format("wl-copy '%s' && notify-send 'Copied image path' -i '%s'", image.path, image.path))
+    local image = swayimg.viewer.get_image()
+    os.execute(string.format("wl-copy '%s' && notify-send 'Copied image path' -i '%s'", image.path, image.path))
 end)
 
--- Set wallpaper with swww and update colors with pywal
+-- Set wallpaper with awww and update colors with pywal
 swayimg.viewer.on_key("Ctrl+w", function()
-	local image = swayimg.viewer.get_image()
-	os.execute(
-		string.format(
-			"swww img '%s' --transition-type wipe --transition-duration 1 & wal -i '%s' && source ~/.cache/wal/colors.sh && hyprctl keyword general:col.active_border 'rgba(${color1:1}ff)'",
-			image.path,
-			image.path
-		)
-	)
+    local image = swayimg.viewer.get_image()
+    os.execute(
+        string.format(
+            "awww img '%s' --transition-type wipe --transition-duration 1 & wal -i '%s' && $HOME/.local/scripts/helpers/refresh_windowborder",
+            image.path,
+            image.path
+        )
+    )
 end)
 
 -- Add image to Favorites
 swayimg.viewer.on_key("Ctrl+f", function()
-	local image = swayimg.viewer.get_image()
-	os.execute(
-		string.format(
-			"cp '%s' ~/Pictures/Favorites/ && notify-send 'Added image to Favorites' -i '%s'",
-			image.path,
-			image.path
-		)
-	)
+    local image = swayimg.viewer.get_image()
+    os.execute(
+        string.format(
+            "cp '%s' ~/Pictures/Favorites/ && notify-send 'Added image to Favorites' -i '%s'",
+            image.path,
+            image.path
+        )
+    )
 end)
 
 -- Delete image
 swayimg.viewer.on_key("Delete", function()
-	local image = swayimg.viewer.get_image()
-	os.execute(string.format("rm '%s' && notify-send 'Image deleted' -i '%s'", image.path, image.path))
-	swayimg.imagelist.remove(image.path)
+    local image = swayimg.viewer.get_image()
+    if confirm("Delete this image?") then
+        os.execute(string.format("rm '%s' && notify-send 'Image deleted' -i '%s'", image.path, image.path))
+        swayimg.imagelist.remove(image.path)
+    end
 end)
